moved methods to interface
